Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -2,8 +2,8 @@
 # include <stdio.h>
 
 int main() {
-    int a, b;
-    scanf("%d %d",&a, &b);
+    float a, b;
+    scanf("%f %f",&a, &b);
     printf("%d", a+b);
     printf("%d", a/b);
     printf("%d", a*b);
